Use 'magic_enum' library for stat enums
