+++ nano/lib/CMakeLists.txt
@@ -68,6 +68,7 @@ add_library(
   signal_manager.cpp
   stats.hpp
   stats.cpp
+  stats_enums.hpp
   stream.hpp
   threading.hpp
   threading.cpp
