+++ nano/node/bootstrap/bootstrap_server.hpp
@@ -29,7 +29,7 @@ public:
 	using request_t = std::pair<nano::asc_pull_req, std::shared_ptr<nano::transport::channel>>; // <request, response channel>
 
 public:
-	bootstrap_server (nano::store &, nano::ledger &, nano::network_constants const &, nano::stat &);
+	bootstrap_server (nano::store &, nano::ledger &, nano::network_constants const &, nano::stats &);
 	~bootstrap_server ();
 
 	void start ();
@@ -74,7 +74,7 @@ private: // Dependencies
 	nano::store & store;
 	nano::ledger & ledger;
 	nano::network_constants const & network_constants;
-	nano::stat & stats;
+	nano::stats & stats;
 
 private:
 	processing_queue<request_t> request_queue;
