+++ nano/node/vote_processor.hpp
@@ -23,7 +23,7 @@ class rep_crawler;
 class ledger;
 class network_params;
 class node_flags;
-class stat;
+class stats;
 
 class transaction;
 namespace transport
@@ -34,14 +34,15 @@ namespace transport
 class vote_processor final
 {
 public:
-	explicit vote_processor (nano::signature_checker & checker_a, nano::active_transactions & active_a, nano::node_observers & observers_a, nano::stat & stats_a, nano::node_config & config_a, nano::node_flags & flags_a, nano::logger_mt & logger_a, nano::online_reps & online_reps_a, nano::rep_crawler & rep_crawler_a, nano::ledger & ledger_a, nano::network_params & network_params_a);
+	vote_processor (nano::signature_checker & checker_a, nano::active_transactions & active_a, nano::node_observers & observers_a, nano::stats & stats_a, nano::node_config & config_a, nano::node_flags & flags_a, nano::logger_mt & logger_a, nano::online_reps & online_reps_a, nano::rep_crawler & rep_crawler_a, nano::ledger & ledger_a, nano::network_params & network_params_a);
+
 	/** Returns false if the vote was processed */
 	bool vote (std::shared_ptr<nano::vote> const &, std::shared_ptr<nano::transport::channel> const &);
 	/** Note: node.active.mutex lock is required */
 	nano::vote_code vote_blocking (std::shared_ptr<nano::vote> const &, std::shared_ptr<nano::transport::channel> const &, bool = false);
 	void verify_votes (std::deque<std::pair<std::shared_ptr<nano::vote>, std::shared_ptr<nano::transport::channel>>> const &);
 	/** Function blocks until either the current queue size (a established flush boundary as it'll continue to increase)
-	  * is processed or the queue is empty (end condition or cutoff's guard, as it is positioned ahead) */
+	 * is processed or the queue is empty (end condition or cutoff's guard, as it is positioned ahead) */
 	void flush ();
 	std::size_t size ();
 	bool empty ();
@@ -56,7 +57,7 @@ private:
 	nano::signature_checker & checker;
 	nano::active_transactions & active;
 	nano::node_observers & observers;
-	nano::stat & stats;
+	nano::stats & stats;
 	nano::node_config & config;
 	nano::logger_mt & logger;
 	nano::online_reps & online_reps;
